chore: extract admin flags routes (#298 4b)
Move GET/POST /admin/merch/flags (217L) from ingest/routes.ts (1228L →
~1010L) into ingest/routes/admin-flags.ts. Pure file move — no behavior
change, no new deps, same auth allowlist and FlagsStore handling.

routes.ts now imports `adminFlagsRoutes(deps)` and spreads it, keeping
createRoutes as the single source. Prepares for further domain splits
without growing the monolith.

Verified: tsc pass, prettier pass, 854 tests pass.
